requirements.txt
numpy>=1.0
pandas>=1.5
pinecone-client>=2.2
pymongo>=3.12
scikit_learn>=1.0
openai>=1.3
python-dotenv>=1.0